+++ driver/command/command_driver.go
@@ -1,4 +1,4 @@
-package driver
+package command
 
 import (
 	"bytes"
