+++ driver/docker/docker_driver.go
@@ -1,4 +1,4 @@
-package driver
+package docker
 
 import (
 	"archive/tar"
