+++ driver/kubernetes/kubernetes.go
@@ -1,4 +1,4 @@
-package driver
+package kubernetes
 
 import (
 	"fmt"
