+++ driver/kubernetes/kubernetes_test.go
@@ -1,4 +1,4 @@
-package driver
+package kubernetes
 
 import (
 	"os"
